refactor(craft): point all callers at canonical session modules
